Converted from hyperscript helpers.
